[UMPA-206] Improve : 접근 제어자 지정
